requirements.txt
requests == 2.24.0
terminaltables == 3.1.10
python-dotenv == 0.21.1s